do 4.21 and fix get blog by id

+++ part4/blog/controllers/blogs.js
@@ -10,7 +10,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.get('/:id', async (request, response) => {
   
-  const blog = await blog.findById(request.params.id)
+  const blog = await Blog.findById(request.params.id)
   if (blog) {
     response.json(blog)
   } else {
@@ -19,8 +19,20 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else {
+    response.status(400).json({ error: 'wrong user or token' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
